STYLE: Reorganized code for better modularity.
