+++ DWIConvert/CMakeLists.txt
@@ -30,6 +30,8 @@ set(DWIConvertSupportLib_SRCS
   DWIConvertUtils.cxx
   DWIConverterFactory.h
   DWIConverter.h
+  DWIDICOMConverterBase.h
+  DWIDICOMConverterBase.cxx
   GEDWIConverter.h
   GenericDWIConverter.h
   HitachiDWIConverter.h
@@ -37,12 +39,6 @@ set(DWIConvertSupportLib_SRCS
   SiemensDWIConverter.h
   StringContains.h
   )
-
-set(DWIConvertTest_SRC)
-foreach(f ${DWIConvertSupportLib_SRCS})
-  list(APPEND DWIConvertTest_SRC
-    ${CMAKE_CURRENT_LIST_DIR}/${f})
-endforeach()
 
 add_library(DWIConvertSupportLib STATIC ${DWIConvertSupportLib_SRCS})
 target_link_libraries(DWIConvertSupportLib ${DWIConvert_ITK_LIBRARIES} BRAINSCommonLib)
@@ -52,7 +48,7 @@ set(PROGS_LIST_1
   )
 foreach(prog ${PROGS_LIST_1})
   add_executable(${prog} ${prog}.cxx)
-  target_link_libraries(${prog} BRAINSCommonLib ${DWIConvert_ITK_LIBRARIES})
+  target_link_libraries(${prog} DWIConvertSupportLib)
 endforeach()
 
 #-----------------------------------------------------------------------------
@@ -65,6 +61,12 @@ endforeach()
 # several files needed down in ExtenededTesting
 StandardBRAINSBuildMacro(NAME DWIConvert EXECUTABLE_ONLY
   TARGET_LIBRARIES DWIConvertSupportLib)
+
+set(DWIConvertTest_SRC)
+foreach(f ${DWIConvertSupportLib_SRCS})
+  list(APPEND DWIConvertTest_SRC
+          ${CMAKE_CURRENT_LIST_DIR}/${f})
+endforeach()
 
 if(BUILD_TESTING AND NOT BRAINSTools_DISABLE_TESTING)
   add_subdirectory(TestSuite)
