+++ DWIConvert/DWIDICOMConverterBase.cxx
@@ -1,0 +1,10 @@
+//
+// Created by Johnson, Hans J on 11/24/16.
+//
+
+#include "DWIDICOMConverterBase.h"
+
+int BLAHBLAHBLAH()
+{
+   return 7;
+}
