+++ DWIConvert/GenericDWIConverter.h
@@ -23,22 +23,25 @@
 class GenericDWIConverter : public DWIConverter
 {
 public:
-  GenericDWIConverter(DWIConverter::DCMTKFileVector &allHeaders,
-                      DWIConverter::FileNamesContainer &inputFileNames,
-                      bool useBMatrixGradientDirections) :  DWIConverter(allHeaders,
-                                                                         inputFileNames,
-                                                                         useBMatrixGradientDirections)
+  GenericDWIConverter( DWIConverter::FileNamesContainer &inputFileNames )
+    :  DWIConverter( inputFileNames)
     {
     }
+
+  virtual void LoadFromDisk() ITK_OVERRIDE
+  {
+    itkGenericExceptionMacro(<< " LoadFromDisk not relevant" << std::endl);
+  }
+
   virtual ~GenericDWIConverter() {}
 protected:
   virtual void ExtractDWIData() ITK_OVERRIDE
     {
-      // throw; // don't call
+        itkGenericExceptionMacro(<< " ExtractDWIData not relevant" << std::endl);
     }
   virtual void AddFlagsToDictionary() ITK_OVERRIDE
     {
-      // throw; // don't call
+        itkGenericExceptionMacro(<< " AddFlagsToDictionary not relevant" << std::endl);
     }
 };
